+++ experiments/notebooks/utils/heatmap_widget/__init__.py
@@ -1,5 +1,18 @@
 # Heatmap Widget Package
 
-from .heatmap_widget import HeatmapWidget, create_demo_heatmap, create_heatmap_widget, create_multi_metric_demo
+from .heatmap_widget.HeatmapWidget import (
+    HeatmapWidget,
+    create_demo_heatmap,
+    create_heatmap_widget,
+    create_multi_metric_demo,
+)
+from .heatmap_widget.metta_client import MettaAPIClient, fetch_real_heatmap_data
 
-__all__ = ["HeatmapWidget", "create_demo_heatmap", "create_heatmap_widget", "create_multi_metric_demo"]
+__all__ = [
+    "HeatmapWidget",
+    "create_demo_heatmap",
+    "create_heatmap_widget",
+    "create_multi_metric_demo",
+    "MettaAPIClient",
+    "fetch_real_heatmap_data",
+]
